Edited

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
 
-    public double calculate(double height, int weight) {
-        return weight / (height * height);
+    public int calculate(double height, int weight) {
+        return (int) (weight / (height * height));
     }
 }
